prov/linkx: Implement getname() callback
The LINKx getname() method collects all the addresses of the different
providers it manages. LINKx can manage multiple providers and each
provider can manage one or more endpoints.

It builds an lnx_addresses structure big enough to fit all the addresses.
If the buffer provided is smaller than that, it sets the addresslen
parameter to the needed size and return FI_ETOOSMALL.

This address defines all the different ways this process can be reached.
For example if a node has 4 NICs, the address can contain the address for
all 4 NICs the process is listening on.

The application, should then allocate a large enough buffer and callback
into the LINKx provider using fi_getname() to get the addresses.

Signed-off-by: Amir Shehata <shehataa@ornl.gov>