hmem/rocr: Support asynchronous memory copies
A key performance improvement for shm provider is to utilize asynchronous
memory copies provided by the device, if available.

This implementation targets ROCR, however, it can be expanded to handle
other devices.

In the smr_progress_ipc() function instead of calling into the synchronous
copy, we call into the asynchronous copy and similar to SAR we return a
structure which will be queued on a local queue. On progress, the queue is
walked and the copy_query HMEM API is called to check when the copy has
completed.

Signed-off-by: Amir Shehata <shehataa@ornl.gov>