prov/linkx: do mem reg within LINKx
If the application didn't do any memory registration, the SHM provider
will not be able to select the protocol to use accurately. SHM provider
relies on the memory registration to determine the type of memory, device
vs system. In contrast other providers use the ofi_get_hmem_iface() API to
determine the iface based on the buffer pointer.

In order to work around that, if the application didn't register the
memory for the SHM provider, then LINKx will find out the buffer type,
device vs system, and do it's own memory registration per send/recv API
call.

Signed-off-by: Amir Shehata <shehataa@ornl.gov>