prov/shm: Change command queue implementation
Instead of using the circular queue for the command queue, use the
freestack for a command block pool. This allows us to pop a command off
the pool and have it not accessible by any other thread even when not using
locking.

Use a circular queue to post command entries. Each command entry contains
two offsets, one for the command block and the other for an RMA command
block if needed, as the RMA protocol requires two commands.

The order of operations is as follows:

Sending
1. pop a command off the command block pool
2. Initialize the command with data needed
3. get the command block offset in the shared region
4. get a command queue entry and initialize the offsets
5. commit the command queue entry on the command queue
6. signal the peer that a command is ready

Receiving
1. iterate through the command entry queue
2. make a local copy of the command entry
3. discard the command entry
4. get a pointer to the command block using the offset
5. process the command
6. push the command block back on the freestack

Signed-off-by: Amir Shehata <shehataa@ornl.org>