prov/shm: Optimize locking
The locking which the SHM provider uses is quiet coarse. It locks the
shared memory region lock across the entire memory operation. This would
force the processes which are sending to the same peer, to serialize their
operations.

Instead, only lock the critical sections which the region lock is intended
to protect. These critical sections are:

  1. The command pool
  2. The command queue
  3. The SAR list

The response queue is protected by the ep->tx_lock.

This patch also changes the lock ordering. If both ep->tx_lock and
region->lock are to be held at the same time, the order of operations
should be:

  - pthread_spin_lock(&ep->tx_lock)
  - pthread_spin_lock(&region->lock)
  - pthread_spin_unlock(&region->lock)
  - pthread_spin_unlock(&ep->tx_lock)

Signed-off-by: Amir Shehata <shehataa@ornl.gov>