prov/shm: Handle XPMEM caching properly
With XPMEM each process maps its address space for other processes to
attach to. When these address spaces are attached to they can overlap.
This presents a problem when caching these addresses in order to avoid
constant xpmem_attach/xpmem_detach, since the address is no longer unique
across all peers.

This patch adds an ID field to the cache key, which is set to the peer id.
This identifies each XPMEM request by the ID of the peer the mapping
belongs to as well as the address and length of the buffer; therefore
resolving the uniqueness problem.

Signed-off-by: Amir Shehata <shehataa@ornl.gov>