prov/shm: Support unaligned xpmem access
XPMEM kernel module has a requirement that addresses and lengths used with
the APIs must be page aligned.

It is safe to xpmem_attach a larger area than the remote buffer size.
xpmem_attach expands the VMA of the process to have enough space to bring
in the remote process pages. It sets up fault handlers to actually bring
in the pages when an access to the virtual address space is done by the
user space process. The call to xpmem_attach(), however, doesn't touch the
pages, just sets up the VMA and internal kernel structure to track the
attached segment.

This patch down aligns the address to the top of the nearest page and up
aligns the end of the buffer to the end of the page it exists in. It then
attaches the larger segment, and accesses the data based on the offset
between the base attached address and the beginning of the buffer being
processed.

Signed-off-by: Amir Shehata <shehataa@ornl.gov>