prov/lnx: dump statistics
Added the ability to dump send statistics. This is important for the
multi-rail feature to be able to view how the different domains are
being used during the run time of the application.

Also keep stats on max and average length of the SRX receive and
unexpected queues. This will aid in characterizing the load on the system.

Signed-off-by: Amir Shehata <shehataa@ornl.gov>